Move database reading/parsing/writing functions to their own files That way we can build a separate binary for rule management if necessary Signed-off-by: Jed <lejosnej@ainfosec.com>
